LayoutTests/http/tests/security/contentSecurityPolicy/WebAssembly-blocked-in-subframe.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
}
</script>
</head>
<body>
<p>Tests that WebAssembly is blocked in a subframe that disallows WebAssembly when the parent frame allows WebAssembly.</p>
<iframe src="WebAssembly-blocked.html"></iframe>
</body>
</html>
